+++ src/analyzers/concept/annalyns-infiltration/index.ts
@@ -3,3 +3,5 @@ import { ExemplarAnalyzer } from '../__exemplar'
 export class AnnalynsInfiltrationAnalyzer extends ExemplarAnalyzer {
   // TODO: implement actual analyzer
 }
+
+export default AnnalynsInfiltrationAnalyzer