+++ src/analyzers/concept/freelancer-rates/index.ts
@@ -3,3 +3,5 @@ import { ExemplarAnalyzer } from '../__exemplar'
 export class FreelancerRatesAnalyzer extends ExemplarAnalyzer {
   // TODO: implement actual analyzer
 }
+
+export default FreelancerRatesAnalyzer