+++ src/analyzers/practice/two-fer/index.ts
@@ -578,3 +578,5 @@ export class TwoFerAnalyzer extends AnalyzerImpl {
     )
   }
 }
+
+export default TwoFerAnalyzer