+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
-## 0.17.0
+## 0.18.0
 
 - Add check for no value returned in `two-fer`
+
+## 0.17.1
+
+- Change the way autoloads work so they are more predicatable
+
+## 0.17.0
+
+- Improve analysis for `lasagna`
 
 ## 0.16.0